requirements.txt
accelerate==1.1.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asyncpg==0.30.0
attrs==24.2.0
backoff==2.2.1
black==24.10.0
boilerpy3==1.0.7
build==1.2.2.post1
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Events==0.5
farm-haystack==1.26.4
fastapi==0.115.4
filelock==3.16.1
fsspec==2024.2.0
greenlet==3.1.1
h11==0.14.0
httpx==0.28.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-imports==0.3.1
Mako==1.3.6
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
num2words==0.5.13
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
pip==24.0
pip-tools==7.4.1
platformdirs==4.3.6
posthog==3.7.0
prompthub==0.1.17
prompthub-py==4.0.0
psutil==6.1.0
psycopg2-binary==2.9.10
pydantic==1.10.19
pydantic_core==2.25.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
quantulum3==0.9.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-cache==0.9.8
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sseclient==0.0.27
sseclient-py==1.8.0
starlette==0.40.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.66.6
transformers==4.46.3
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.44.0
